Major overhaul to the system.
- Removed FPV_Collection
- Removed FPV_Container

- System no longer uses a collection of all FPV_Objects to iterate and enable/disable first person view.

To simplify, here’s the new process in a sequential way:

On Update
FPV_DisableOnly will check if a renderer inside was changed by the FPV. If it was, it means that the renderers were disabled for the First Person View, so in this Update, the method will re-enable the renderers so they are active when the World Camera is going to render.

World Camera - OnWillRenderObject
FPV_Renderer will first check if it is not a first person object. Since the World Camera is currently rendering, it will check if the renderer was changed by the First Person Camera. If it was, it will revert it back to normal so the World camera will render it correctly
FPV_Renderer_DisableOnly will first check if it is not a first person object. It will then check if the world camera is rendering and will enable the First Person View for this renderer to prepare it for the First Person Camera.

First Person Camera - OnWillRenderObject
FPV_Renderer will first check if it is not a first person object. Since the First Person Camera is rendering, it will enable the First Person View of the renderer ( in this case, it will set the shadow casting mode to ShadowsOnly.